+++ .github/CHANGELOG.md
@@ -31,6 +31,7 @@ This is the changelog for the version V2.6.1. Don't forget to [join us on Discor
 
 * Use Statevents to increase Rush-Stability
 * Bots use Squad-Perks
+* First implementation of Alternate Weapons on some vehicles
 
 ### Bug fixes
 * #218 increased auto-sniper-time a little